fix: offscreen worker file struct in wxt framework
